Homework1

+++ src/Homework1.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Homework1 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextDouble();
-        double b = sc.nextDouble();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
 
-        System.out.println((int)(a + b));
+        System.out.println(a + b);
     }
 }
